init builder pattern, refactor game_config into models
