+++ lib/presentation/game_room/builder/director.dart
@@ -1,0 +1,11 @@
+import 'package:chess_game/presentation/game_room/builder/interface/game_config_builder.dart';
+
+// abstract class Director {
+//   IGameBuilder? builder;
+
+//   void setBuilder(IGameBuilder builder) {
+//     this.builder = builder;
+//   }
+  
+//   void buildGameConfig() {}
+// }
