+++ lib/presentation/game_room/builder/interface/game_config_builder.dart
@@ -1,0 +1,14 @@
+import 'package:chess_game/core/models/game_config.dart';
+
+// Builder pattern
+
+abstract class IGameConfigBuilder {
+  IGameConfigBuilder reset();
+  IGameConfigBuilder setTimeControl(int minutes, int seconds);
+  IGameConfigBuilder setPlayerType(String playerType);
+  IGameConfigBuilder setBoardTheme(String boardTheme);
+  IGameConfigBuilder setPieceSet(String pieceSet);
+  IGameConfigBuilder setSoundEnabled(bool soundEnabled);
+  IGameConfigBuilder setDifficultyLevel(int difficultyLevel);
+  GameConfig build();
+}
